move non-article specific parts into CollectionViewCell
